+++ src/main/java/com/craftblocklock/client/LockedRecipeWidget.java
@@ -1,0 +1,5 @@
+package com.craftblocklock.client;
+
+public interface LockedRecipeWidget {
+    boolean craftblocklock$isLockedRecipe();
+}